Move MMREventEmitter to server utils
